cukinia_stat: add new function to check file ownership, and permissions
The cukinia_test command was useful to test file existence, and
permissions, but only for the user running cukinia. The new
cukinia_stat function allows to check file existence, ownership
(user and group), and permissions (mode) in a more complete way.

It may be used to ensure that sensitive files (like /etc/shadow)
have the correct ownership and permissions even regarding other
users.